refactor(review): 리뷰 반영
