+++ src/main/kotlin/lotto/Runner.kt
@@ -1,7 +1,16 @@
 package lotto
 
-import lotto.domain.LottoGame
+import lotto.domain.LottoMachine
+import lotto.domain.LottoMachineResult
+import lotto.view.InputView
+import lotto.view.OutputView
 
 fun main(args: Array<String>) {
-    LottoGame.runGame()
+    val purchaseInfo = InputView.makePurchaseInfo()
+    val userLottoTickets = LottoMachine(purchaseInfo).makeLottoTickets()
+    OutputView.printPurchasingLottoNumber(userLottoTickets)
+
+    val winLottoInfo = InputView.makeWinLottoInfo()
+    val lottoMachineResult = LottoMachineResult(userLottoTickets, winLottoInfo)
+    OutputView.printWinnerStatistics(lottoMachineResult, purchaseInfo)
 }
