+++ src/main/kotlin/lotto/view/InputView.kt
@@ -2,7 +2,6 @@ package lotto.view
 
 import lotto.supportdata.PurchaseInfo
 import lotto.supportdata.WinLottoInfo
-import java.lang.IllegalArgumentException
 
 object InputView {
     private tailrec fun purchase(): Int {
@@ -25,11 +24,6 @@ object InputView {
     fun makeWinLottoInfo(): WinLottoInfo {
         val winNumber = winNumbers()
         val bonusBall = bonusBall()
-        return try {
-            WinLottoInfo(winNumber, bonusBall)
-        } catch (e: IllegalArgumentException) {
-            println("다시 입력해주세요 \nㄴ사유 : ${e.message} \n")
-            makeWinLottoInfo()
-        }
+        return WinLottoInfo(winNumber, bonusBall)
     }
 }
