+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -19,7 +19,11 @@ object OutputView {
         println("---------")
         LottoRank.values()
             .filter { it != LottoRank.OUT }
-            .forEach { println("${it.infoString} - ${lottoMachineResult.getLottoRankCount(it)}개") }
+            .forEach { println("${it.rankInfoString()} - ${lottoMachineResult.getLottoRankCount(it)}개") }
         println("총 수익률은 ${lottoMachineResult.calculateProfit(purchaseInfo)}입니다.")
     }
+
+    private fun LottoRank.rankInfoString(): String =
+        if (this == LottoRank.BONUS) "${matchCount}개 일치, 보너스 볼 일치(${price}원)"
+        else "${matchCount}개 일치 (${price}원)"
 }
